:chore: :children_crossing: updatedexamples
refactoring of examples to give them better names and explaining comments
