+++ examples/module/src/main/java/module/ExampleModule.java
@@ -1,0 +1,11 @@
+package module;
+
+import skillup.annotations.Module;
+
+/**
+ * An module example which has to be provided with annotation @Module
+ */
+@Module(description = "A simple example module that acts as a container but does not provide any more information", moduleIri = "https://hsu-hh.de/modules#ExampleModule")
+public class ExampleModule {
+
+}
