--- examples/module/src/main/java/moduleA/ModuleA.java
@@ -1,11 +1,0 @@
-package moduleA;
-
-import skillup.annotations.Module;
-
-/**
- * An module example which has to be provided with annotation @Module
- */
-@Module(description = "MyModule", moduleIri = "https://hsu-hh.de/modules#ModuleA")
-public class ModuleA {
-
-}
