--- examples/opc-ua-skill/pom.xml
@@ -1,49 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>io.github.aljoshakoecher.skillup</groupId>
-		<artifactId>examples</artifactId>
-		<version>${revision}</version>
-	</parent>
-	
-	<artifactId>skillup.examples.opc-ua-skill</artifactId>
-	<packaging>bundle</packaging>
-	<name>OPC-UA-Skill</name>
-	
-	<dependencies>
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-simple</artifactId>
-			<version>1.7.28</version>
-		</dependency>
-		<dependency>
-			<groupId>io.github.aljoshakoecher.skillup</groupId>
-			<artifactId>skillup.annotations</artifactId>
-			<version>${revision}</version>
-		</dependency>
-	</dependencies>
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.8.0</version>
-				<configuration>
-					<release>11</release>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.felix</groupId>
-				<artifactId>maven-bundle-plugin</artifactId>
-				<extensions>true</extensions>
-				<configuration>
-					<instructions>
-						<_dsannotations>*</_dsannotations>
-					</instructions>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
