+++ examples/opcua-add-skill/.gitignore
