+++ examples/opcua-random-generator/pom.xml
