+++ examples/rest-doubler/.gitignore
