+++ examples/rest-doubler/pom.xml
