+++ examples/rest-multiplier/pom.xml
